Workaround: Re-Init SCTimer when duty cycle is set to 0 / turned off
When the duty cycle for one PWM output is set to 0 the NXP driver turns off all PWM outputs.
However, the registers are not cleaned properly.
When turning on another PWM output the first one turns again on with it with weird values.
Implemented quickfix to re-init SCTimer each time the duty cycle of a PWM output is set to 0.
The clears all registers / parameters to start from a fresh start.
Caviat: Cannot turn off individual PWM outputs directly